Move crypto to a dir
