+++ src/header.rs
@@ -2,6 +2,8 @@ use serde::{Deserialize, Serialize};
 
 use crate::algorithms::Algorithm;
 use crate::errors::Result;
+use crate::serialization::decode;
+
 
 /// A basic JWT header, the alg defaults to HS256 and typ is automatically
 /// set to `JWT`. All the other fields are optional.
@@ -59,7 +61,7 @@ impl Header {
 
     /// Converts an encoded part into the Header struct if possible
     pub(crate) fn from_encoded(encoded_part: &str) -> Result<Self> {
-        let decoded = base64::decode_config(encoded_part, base64::URL_SAFE_NO_PAD)?;
+        let decoded = decode(encoded_part)?;
         let s = String::from_utf8(decoded)?;
 
         Ok(serde_json::from_str(&s)?)
